# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Service Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus];

export const TicketStatus = {
  received: "received",
  diagnosing: "diagnosing",
  repairing: "repairing",
  waiting_for_parts: "waiting_for_parts",
  ready_for_pickup: "ready_for_pickup",
  delivered: "delivered",
} as const;

export interface Ticket {
  id: number;
  serviceId: string;
  customerName: string;
  customerPhone: string;
  customerEmail?: string | null;
  deviceBrand: string;
  deviceModel: string;
  problemDescription: string;
  status: TicketStatus;
  technicianId?: number | null;
  technicianName?: string | null;
  estimatedCompletion?: string | null;
  workSummary?: string | null;
  totalPartsAmount?: number | null;
  totalLaborAmount?: number | null;
  totalAmount?: number | null;
  billConfirmed: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface TicketPart {
  id: number;
  ticketId: number;
  partId?: number | null;
  partName: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
  warrantyPeriod?: string | null;
  addedByName?: string | null;
  createdAt: string;
}

export interface LaborEntry {
  id: number;
  ticketId: number;
  technicianId?: number | null;
  technicianName?: string | null;
  startTime?: string | null;
  endTime?: string | null;
  manualHours?: number | null;
  totalHours?: number | null;
  laborCost?: number | null;
  isRunning: boolean;
  createdAt: string;
}

export interface AuditEntry {
  id: number;
  ticketId: number;
  action: string;
  description: string;
  technicianId?: number | null;
  technicianName?: string | null;
  metadata?: string | null;
  createdAt: string;
}

export type TicketDetail = Ticket & {
  parts: TicketPart[];
  laborEntries: LaborEntry[];
  auditLog: AuditEntry[];
};

export interface PublicStatus {
  serviceId: string;
  customerName: string;
  deviceBrand: string;
  deviceModel: string;
  status: TicketStatus;
  statusLabel: string;
  estimatedCompletion?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface Part {
  id: number;
  name: string;
  barcode?: string | null;
  defaultPrice: number;
  stockQuantity: number;
  category?: string | null;
  createdAt: string;
}

export type TechnicianRole =
  (typeof TechnicianRole)[keyof typeof TechnicianRole];

export const TechnicianRole = {
  technician: "technician",
  manager: "manager",
} as const;

export interface Technician {
  id: number;
  name: string;
  pin: string;
  role: TechnicianRole;
  isActive: boolean;
  createdAt: string;
}

export interface Settings {
  id: number;
  hourlyRate: number;
  vatRate: number;
  twilioAccountSid?: string | null;
  twilioAuthToken?: string | null;
  twilioFromNumber?: string | null;
  emailHost?: string | null;
  emailPort?: number | null;
  emailUser?: string | null;
  emailPass?: string | null;
  emailFrom?: string | null;
  printerIp?: string | null;
  printerPort?: number | null;
  smsTemplate?: string | null;
  emailTemplate?: string | null;
  updatedAt: string;
}

export type AnalyticsRevenueByDayItem = {
  date: string;
  revenue: number;
};

export type AnalyticsTopFailureTypesItem = {
  deviceBrand: string;
  deviceModel: string;
  count: number;
};

export type AnalyticsTechnicianStatsItem = {
  technicianName: string;
  jobsCompleted: number;
  avgHoursPerJob: number;
};

export interface Analytics {
  totalRevenue: number;
  totalTickets: number;
  completedTickets: number;
  averageRepairTime: number;
  revenueByDay: AnalyticsRevenueByDayItem[];
  topFailureTypes: AnalyticsTopFailureTypesItem[];
  technicianStats: AnalyticsTechnicianStatsItem[];
}

export interface CreateTicketBody {
  customerName: string;
  customerPhone: string;
  customerEmail?: string | null;
  deviceBrand: string;
  deviceModel: string;
  problemDescription: string;
  technicianId?: number | null;
  estimatedCompletion?: string | null;
}

export interface UpdateTicketBody {
  customerName?: string;
  customerPhone?: string;
  customerEmail?: string | null;
  deviceBrand?: string;
  deviceModel?: string;
  problemDescription?: string;
  technicianId?: number | null;
  estimatedCompletion?: string | null;
}

export interface UpdateStatusBody {
  status: TicketStatus;
  technicianId?: number | null;
  reason?: string | null;
  billConfirmed?: boolean | null;
}

export interface AddPartBody {
  partId?: number | null;
  partName: string;
  quantity: number;
  unitPrice: number;
  warrantyPeriod?: string | null;
  technicianId?: number | null;
}

export type AddLaborBodyMode =
  (typeof AddLaborBodyMode)[keyof typeof AddLaborBodyMode];

export const AddLaborBodyMode = {
  timer: "timer",
  manual: "manual",
} as const;

export interface AddLaborBody {
  technicianId?: number | null;
  mode: AddLaborBodyMode;
  manualHours?: number | null;
  hourlyRate?: number | null;
}

export interface WorkSummaryBody {
  workSummary: string;
  technicianId?: number | null;
}

export interface AddNoteBody {
  note: string;
  technicianId?: number | null;
}

export interface CreatePartBody {
  name: string;
  barcode?: string | null;
  defaultPrice: number;
  stockQuantity: number;
  category?: string | null;
}

export type CreateTechnicianBodyRole =
  (typeof CreateTechnicianBodyRole)[keyof typeof CreateTechnicianBodyRole];

export const CreateTechnicianBodyRole = {
  technician: "technician",
  manager: "manager",
} as const;

export interface CreateTechnicianBody {
  name: string;
  pin: string;
  role: CreateTechnicianBodyRole;
}

export interface LoginBody {
  pin: string;
}

export interface UpdateSettingsBody {
  hourlyRate?: number;
  vatRate?: number;
  twilioAccountSid?: string | null;
  twilioAuthToken?: string | null;
  twilioFromNumber?: string | null;
  emailHost?: string | null;
  emailPort?: number | null;
  emailUser?: string | null;
  emailPass?: string | null;
  emailFrom?: string | null;
  printerIp?: string | null;
  printerPort?: number | null;
  smsTemplate?: string | null;
  emailTemplate?: string | null;
}

export type ListTicketsParams = {
  status?: TicketStatus;
  technicianId?: number;
  search?: string;
  dateFrom?: string;
  dateTo?: string;
};

export type ListPartsParams = {
  search?: string;
  barcode?: string;
};

export type GetAnalyticsParams = {
  period?: GetAnalyticsPeriod;
};

export type GetAnalyticsPeriod =
  (typeof GetAnalyticsPeriod)[keyof typeof GetAnalyticsPeriod];

export const GetAnalyticsPeriod = {
  week: "week",
  month: "month",
  year: "year",
} as const;

export type SeedData200 = {
  message: string;
};