# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Service Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /parts:
    get:
      operationId: listParts
      tags: [parts]
      summary: List inventory parts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: barcode
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Parts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Part"

    post:
      operationId: createPart
      tags: [parts]
      summary: Create a new inventory part
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePartBody"
      responses:
        "201":
          description: Part created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Part"
  /parts/{id}:
    patch:
      operationId: updatePart
      tags: [parts]
      summary: Update an inventory part
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePartBody"
      responses:
        "200":
          description: Part updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Part"

    delete:
      operationId: deletePart
      tags: [parts]
      summary: Delete an inventory part
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Part deleted
  /technicians:
    get:
      operationId: listTechnicians
      tags: [technicians]
      summary: List all technicians
      responses:
        "200":
          description: Technicians list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Technician"

    post:
      operationId: createTechnician
      tags: [technicians]
      summary: Create a technician
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTechnicianBody"
      responses:
        "201":
          description: Technician created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
  /technicians/login:
    post:
      operationId: technicianLogin
      tags: [technicians]
      summary: Technician PIN login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get admin settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"

    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update admin settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsBody"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
  /analytics/dashboard:
    get:
      operationId: getAnalytics
      tags: [settings]
      summary: Get analytics dashboard data
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year]
            default: month
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analytics"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Part:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        barcode:
          type: string
          nullable: true
        defaultPrice:
          type: number
        stockQuantity:
          type: integer
        category:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - defaultPrice
        - stockQuantity
        - createdAt
    Technician:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        pin:
          type: string
        role:
          type: string
          enum: [technician, manager]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - pin
        - role
        - isActive
        - createdAt
    Settings:
      type: object
      properties:
        id:
          type: integer
        hourlyRate:
          type: number
        vatRate:
          type: number
        twilioAccountSid:
          type: string
          nullable: true
        twilioAuthToken:
          type: string
          nullable: true
        twilioFromNumber:
          type: string
          nullable: true
        emailHost:
          type: string
          nullable: true
        emailPort:
          type: integer
          nullable: true
        emailUser:
          type: string
          nullable: true
        emailPass:
          type: string
          nullable: true
        emailFrom:
          type: string
          nullable: true
        printerIp:
          type: string
          nullable: true
        printerPort:
          type: integer
          nullable: true
        smsTemplate:
          type: string
          nullable: true
        emailTemplate:
          type: string
          nullable: true
        updatedAt:
          type: string
      required:
        - id
        - hourlyRate
        - vatRate
        - updatedAt
    Analytics:
      type: object
      properties:
        totalRevenue:
          type: number
        totalTickets:
          type: integer
        completedTickets:
          type: integer
        averageRepairTime:
          type: number
        revenueByDay:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              revenue:
                type: number
            required:
              - date
              - revenue
        topFailureTypes:
          type: array
          items:
            type: object
            properties:
              deviceBrand:
                type: string
              deviceModel:
                type: string
              count:
                type: integer
            required:
              - deviceBrand
              - deviceModel
              - count
        technicianStats:
          type: array
          items:
            type: object
            properties:
              technicianName:
                type: string
              jobsCompleted:
                type: integer
              avgHoursPerJob:
                type: number
            required:
              - technicianName
              - jobsCompleted
              - avgHoursPerJob
      required:
        - totalRevenue
        - totalTickets
        - completedTickets
        - averageRepairTime
        - revenueByDay
        - topFailureTypes
        - technicianStats
...
    CreatePartBody:
      type: object
      properties:
        name:
          type: string
        barcode:
          type: string
          nullable: true
        defaultPrice:
          type: number
        stockQuantity:
          type: integer
        category:
          type: string
          nullable: true
      required:
        - name
        - defaultPrice
        - stockQuantity
    CreateTechnicianBody:
      type: object
      properties:
        name:
          type: string
        pin:
          type: string
        role:
          type: string
          enum: [technician, manager]
      required:
        - name
        - pin
        - role
    LoginBody:
      type: object
      properties:
        pin:
          type: string
      required:
        - pin
    UpdateSettingsBody:
      type: object
      properties:
        hourlyRate:
          type: number
        vatRate:
          type: number
        twilioAccountSid:
          type: string
          nullable: true
        twilioAuthToken:
          type: string
          nullable: true
        twilioFromNumber:
          type: string
          nullable: true
        emailHost:
          type: string
          nullable: true
        emailPort:
          type: integer
          nullable: true
        emailUser:
          type: string
          nullable: true
        emailPass:
          type: string
          nullable: true
        emailFrom:
          type: string
          nullable: true
        printerIp:
          type: string
          nullable: true
        printerPort:
          type: integer
          nullable: true
        smsTemplate:
          type: string
          nullable: true
        emailTemplate:
          type: string
          nullable: true