# lib/api-zod/src/generated/types/analytics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Service Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyticsRevenueByDayItem } from "./analyticsRevenueByDayItem";
import type { AnalyticsTechnicianStatsItem } from "./analyticsTechnicianStatsItem";
import type { AnalyticsTopFailureTypesItem } from "./analyticsTopFailureTypesItem";

export interface Analytics {
  totalRevenue: number;
  totalTickets: number;
  completedTickets: number;
  averageRepairTime: number;
  revenueByDay: AnalyticsRevenueByDayItem[];
  topFailureTypes: AnalyticsTopFailureTypesItem[];
  technicianStats: AnalyticsTechnicianStatsItem[];
}